% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbump.R
\name{geom_bump}
\alias{geom_bump}
\title{Smooth rank over time}
\usage{
geom_bump(
  mapping = NULL,
  data = NULL,
  geom = "line",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\value{
ggplot layer
}
\description{
Creates a ggplot that makes a smooth rank over time.
}
\examples{
\dontrun{
  df <- tibble(country = c(
  "India", "India", "India",
  "Sweden", "Sweden", "Sweden",
  "Germany", "Germany", "Germany",
  "Finland", "Finland", "Finland"),
year = c(2011, 2012, 2013,
2011, 2012, 2013,
2011, 2012, 2013,
2011, 2012, 2013),
rank = c(4, 2, 2, 3, 1, 4, 2, 3, 1, 1, 4, 3))

ggplot(df, aes(year, rank, color = country)) +
  geom_point(size = 10) +
  geom_bump(size = 2)
}

}
