% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_bump.R, R/geom_sigmoid.R
\docType{data}
\name{ggbump-ggproto}
\alias{ggbump-ggproto}
\alias{StatBump}
\alias{StatSigmoid}
\title{ggbump extensions to ggplot2}
\description{
These ggproto objects are participate in the ggproto class extension
mechanism provided by ggplot2. They are typically of no concern, as they can
be used through the `geom_*()` functions.
}
\keyword{datasets}
