% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbump.R
\name{rank_sigmoid}
\alias{rank_sigmoid}
\title{rank_sigmoid

Creates a longer dataframe with coordinates for a smoothed line.}
\usage{
rank_sigmoid(x, y, smooth = 8, direction = "x")
}
\arguments{
\item{x}{vector}

\item{y}{vector}

\item{smooth}{smooth parameter. Higher means less smoothing}

\item{direction}{the character x or y depending of smoothing direction}
}
\value{
a data frame
}
\description{
rank_sigmoid

Creates a longer dataframe with coordinates for a smoothed line.
}
