% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggbump.R
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid

Creates a longer dataframe with coordinates for a smoothed line.}
\usage{
sigmoid(x_from, x_to, y_from, y_to, smooth = 5, n = 100)
}
\arguments{
\item{x_from}{start x value}

\item{x_to}{end x value}

\item{y_from}{start y value}

\item{y_to}{end y values}

\item{smooth}{smooth parameter. Higher means less smoothing}

\item{n}{number of point that should be smoothed}
}
\value{
a data frame
}
\description{
sigmoid

Creates a longer dataframe with coordinates for a smoothed line.
}
