% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sigmoid}
\alias{sigmoid}
\title{sigmoid}
\usage{
sigmoid(x_from, x_to, y_from, y_to, smooth = 5, n = 100, direction = "x")
}
\arguments{
\item{x_from}{start x value}

\item{x_to}{end x value}

\item{y_from}{start y value}

\item{y_to}{end y values}

\item{smooth}{smooth parameter. Higher means less smoothing}

\item{n}{number of point that should be smoothed}

\item{direction}{the character x or y depending on direction of smoothing}
}
\value{
a data frame
}
\description{
Creates a longer dataframe with coordinates for a smoothed line.
}
